# src/main/resources/openapi/CV_Management_API_OpenAPI_v1.1.yaml
openapi: 3.1.1
info:
  title: CV Management and Deterministic Internship Candidate Filtering API
  version: 1.1.0
  description: REST/JSON API contract for the locked Version 1 scope, revised for Sprint 2 Complete 
    Authentication for Student and predefined Department Admin accounts.
  x-revision-summary:
    - Expanded Sprint 2 authentication to predefined Department Admin accounts.
    - Generalized OTP-based password reset to STUDENT and ADMIN account types.
    - Added password-reset OTP resend endpoint.
    - Standardized authentication response DTOs and success status codes.
    - Removed required logout request body.
    - Defined non-enumerating password-reset start behavior.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /student-verifications:
    post:
      tags:
        - Student Verification and OTP
      summary: Start Student sign-up auto-verification using full name, index number and university email.
      operationId: postStudentVerifications
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentVerificationResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
      description: 'Access: Public. Audit: Security. Full name is display data only; index number and university
        email are verification authority.'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentVerificationStartRequest'
  /student-verifications/{verificationId}/otp/verify:
    post:
      tags:
        - Student Verification and OTP
      summary: Verify the OTP issued after successful student auto-verification.
      operationId: postStudentVerificationsByVerificationIdOtpVerify
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpVerifyResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
      description: 'Access: Public. Audit: Security. Raw OTP is never returned or logged.'
      security: []
      parameters:
        - name: verificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: verificationId identifier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyRequest'
  /student-verifications/{verificationId}/otp/resend:
    post:
      tags:
        - Student Verification and OTP
      summary: Resend a valid OTP after expiry or non-delivery subject to rate limits.
      operationId: postStudentVerificationsByVerificationIdOtpResend
      responses:
        '202':
          description: OTP resend request accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpResendResponse'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
      description: 'Access: Public. Audit: Security. Enforces expiry, resend cooldown, resend limits, and anti-abuse
        controls. Raw OTP values are never returned or logged.'
      security: []
      parameters:
        - name: verificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: verificationId identifier.
  /student-verifications/{verificationId}/password:
    post:
      tags:
        - Student Verification and OTP
      summary: Create initial Student password after OTP verification.
      operationId: postStudentVerificationsByVerificationIdPassword
      responses:
        '204':
          description: Initial Student password created successfully; no response body.
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '415':
          $ref: '#/components/responses/Error415'
        '422':
          $ref: '#/components/responses/Error422'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
      description: 'Access: Public with a verified onboarding context. Audit: Security. Creates the initial Student
        password hash only; no temporary password is generated or returned.'
      security: []
      parameters:
        - name: verificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: verificationId identifier.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordCreateRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error400:
      description: Bad request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error401:
      description: Authentication required or invalid token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error403:
      description: Access denied.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error404:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error409:
      description: Conflict.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error412:
      description: Precondition failed or version conflict.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error415:
      description: Unsupported media type.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error422:
      description: Validation failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error429:
      description: Too many requests.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error500:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Error503:
      description: Service unavailable.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    FieldError:
      type: object
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
...
    OtpResendResponse:
      type: object
      additionalProperties: false
      required:
        - message
        - expiresInSeconds
      properties:
        message:
          type: string
        expiresInSeconds:
          type: integer
          minimum: 1
    OtpVerifyRequest:
      type: object
      required:
        - otpCode
      properties:
        otpCode:
          type: string
          minLength: 6
          maxLength: 6
          writeOnly: true
    OtpVerifyResponse:
      type: object
      additionalProperties: false
      required:
        - verified
      properties:
        verified:
          type: boolean
          const: true
...
    PasswordCreateRequest:
      type: object
      required:
        - newPassword
        - confirmPassword
      properties:
        newPassword:
          type: string
          format: password
          writeOnly: true
        confirmPassword:
          type: string
          format: password
          writeOnly: true
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
        - code
        - message
        - correlationId
      properties:
        type:
          type: string
          format: uri
          examples:
            - https://uor-cv-system/errors/validation-failed
        title:
          type: string
          examples:
            - Validation failed
        status:
          type: integer
          examples:
            - 422
        code:
          type: string
          examples:
            - VALIDATION_FAILED
        message:
          type: string
          examples:
            - One or more fields are invalid.
        correlationId:
          type: string
          examples:
            - req-20260630-0001
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        details:
          type: object
          additionalProperties: true
...
    StudentVerificationResponse:
      type: object
      properties:
        verificationId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - OTP_SENT
            - VERIFICATION_FAILED
        message:
          type: string
        expiresAt:
          type: string
          format: date-time
    StudentVerificationStartRequest:
      type: object
      required:
        - fullName
        - indexNumber
        - universityEmail
      properties:
        fullName:
          type: string
        indexNumber:
          type: string
        universityEmail:
          type: string
          format: email
...